feat: tighten testing, and frame hygiene
