+++ crates/bevy_mod_scripting_core/assets/test_assets/test_script.lua
@@ -1,0 +1,1 @@
+hello world
