+++ crates/bevy_mod_scripting_core/assets/test_assets/test_script.script
@@ -1,0 +1,1 @@
+test script
