+++ crates/bevy_mod_scripting_core/src/context.rs
@@ -11,6 +11,7 @@ impl<T: 'static> Context for T {}
 
 pub type ContextId = u32;
 
+/// Stores script state for a scripting plugin. Scripts are identified by their `ScriptId`, while contexts are identified by their `ContextId`.
 #[derive(Resource)]
 pub struct ScriptContexts<P: IntoScriptPluginParams> {
     pub contexts: HashMap<ContextId, P::C>,
@@ -26,12 +27,6 @@ impl<P: IntoScriptPluginParams> Default for ScriptContexts<P> {
 
 static CONTEXT_ID_COUNTER: AtomicU32 = AtomicU32::new(0);
 impl<P: IntoScriptPluginParams> ScriptContexts<P> {
-    pub fn new() -> Self {
-        Self {
-            contexts: HashMap::new(),
-        }
-    }
-
     /// Allocates a new ContextId and inserts the context into the map
     pub fn insert(&mut self, ctxt: P::C) -> ContextId {
         let id = CONTEXT_ID_COUNTER.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
@@ -64,30 +59,20 @@ pub type ContextInitializer<P> =
 pub type ContextPreHandlingInitializer<P> =
     fn(&str, Entity, &mut <P as IntoScriptPluginParams>::C) -> Result<(), ScriptError>;
 
+/// Settings concerning the creation and assignment of script contexts as well as their initialization.
 #[derive(Resource)]
 pub struct ContextLoadingSettings<P: IntoScriptPluginParams> {
     /// Defines the strategy used to load and reload contexts
-    pub loader: Option<ContextBuilder<P>>,
+    pub loader: ContextBuilder<P>,
     /// Defines the strategy used to assign contexts to scripts
-    pub assigner: Option<ContextAssigner<P>>,
+    pub assigner: ContextAssigner<P>,
     /// Initializers run once after creating a context but before executing it for the first time
     pub context_initializers: Vec<ContextInitializer<P>>,
     /// Initializers run every time before executing or loading a script
     pub context_pre_handling_initializers: Vec<ContextPreHandlingInitializer<P>>,
 }
 
-impl<P: IntoScriptPluginParams> Default for ContextLoadingSettings<P> {
-    fn default() -> Self {
-        Self {
-            loader: None,
-            assigner: None,
-            context_initializers: Default::default(),
-            context_pre_handling_initializers: Default::default(),
-        }
-    }
-}
-
-impl<P: IntoScriptPluginParams> Clone for ContextLoadingSettings<P> {
+impl<T: IntoScriptPluginParams> Clone for ContextLoadingSettings<T> {
     fn clone(&self) -> Self {
         Self {
             loader: self.loader.clone(),
@@ -160,5 +145,56 @@ impl<P: IntoScriptPluginParams> Clone for ContextAssigner<P> {
             assign: self.assign,
             remove: self.remove,
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::asset::Language;
+
+    use super::*;
+
+    struct DummyParams;
+    impl IntoScriptPluginParams for DummyParams {
+        type C = String;
+        type R = ();
+
+        const LANGUAGE: Language = Language::Lua;
+
+        fn build_runtime() -> Self::R {
+            todo!()
+        }
+    }
+
+    #[test]
+    fn test_script_contexts_insert_get() {
+        let mut contexts: ScriptContexts<DummyParams> = ScriptContexts::default();
+        let id = contexts.insert("context1".to_string());
+        assert_eq!(contexts.contexts.get(&id), Some(&"context1".to_string()));
+        assert_eq!(
+            contexts.contexts.get_mut(&id),
+            Some(&mut "context1".to_string())
+        );
+    }
+
+    #[test]
+    fn test_script_contexts_allocate_id() {
+        let contexts: ScriptContexts<DummyParams> = ScriptContexts::default();
+        let id = contexts.allocate_id();
+        let next_id = contexts.allocate_id();
+        assert_eq!(next_id, id + 1);
+    }
+
+    #[test]
+    fn test_script_contexts_remove() {
+        let mut contexts: ScriptContexts<DummyParams> = ScriptContexts::default();
+        let id = contexts.insert("context1".to_string());
+        let removed = contexts.remove(id);
+        assert_eq!(removed, Some("context1".to_string()));
+        assert!(!contexts.contexts.contains_key(&id));
+
+        // assert next id is still incremented
+        let next_id = contexts.allocate_id();
+        assert_eq!(next_id, id + 1);
     }
 }
