+++ crates/bevy_mod_scripting_core/src/lib.rs
@@ -1,13 +1,13 @@
 use crate::event::ScriptErrorEvent;
 use asset::{
-    AssetPathToLanguageMapper, Language, ScriptAsset, ScriptAssetLoader, ScriptAssetSettings,
-    ScriptMetadataStore,
+    configure_asset_systems, configure_asset_systems_for_plugin, AssetPathToLanguageMapper,
+    Language, ScriptAsset, ScriptAssetLoader, ScriptAssetSettings,
 };
 use bevy::prelude::*;
 use bindings::{
-    function::script_function::AppScriptFunctionRegistry, script_value::ScriptValue,
-    AppReflectAllocator, ReflectAllocator, ReflectReference, ScriptTypeRegistration,
-    WorldCallbackAccess,
+    function::script_function::AppScriptFunctionRegistry, garbage_collector,
+    script_value::ScriptValue, AppReflectAllocator, ReflectAllocator, ReflectReference,
+    ScriptTypeRegistration, WorldCallbackAccess,
 };
 use context::{
     Context, ContextAssigner, ContextBuilder, ContextInitializer, ContextLoadingSettings,
@@ -16,13 +16,8 @@ use context::{
 use docs::{Documentation, DocumentationFragment};
 use event::ScriptCallbackEvent;
 use handler::{CallbackSettings, HandlerFn};
-
-use runtime::{Runtime, RuntimeContainer, RuntimeInitializer, RuntimeSettings};
+use runtime::{initialize_runtime, Runtime, RuntimeContainer, RuntimeInitializer, RuntimeSettings};
 use script::Scripts;
-use systems::{
-    garbage_collector, initialize_runtime, insert_script_metadata, remove_script_metadata,
-    sync_script_data, ScriptingSystemSet,
-};
 
 pub mod asset;
 pub mod bindings;
@@ -35,7 +30,20 @@ pub mod handler;
 pub mod reflection_extensions;
 pub mod runtime;
 pub mod script;
-pub mod systems;
+
+#[derive(SystemSet, Hash, Debug, Eq, PartialEq, Clone)]
+/// Labels for various BMS systems
+pub enum ScriptingSystemSet {
+    // Post Setup processes
+    RuntimeInitialization,
+
+    // Post Update processes
+    GarbageCollection,
+
+    ScriptMetadataInsertion,
+    ScriptCommandDispatch,
+    ScriptMetadataRemoval,
+}
 
 /// Types which act like scripting plugins, by selecting a context and runtime
 /// Each individual combination of context and runtime has specific infrastructure built for it and does not interact with other scripting plugins
@@ -56,32 +64,15 @@ pub struct ScriptingPlugin<P: IntoScriptPluginParams> {
     /// The handler used for executing callbacks in scripts
     pub callback_handler: Option<HandlerFn<P>>,
     /// The context builder for loading contexts
-    pub context_builder: Option<ContextBuilder<P>>,
+    pub context_builder: ContextBuilder<P>,
     /// The context assigner for assigning contexts to scripts, if not provided default strategy of keeping each script in its own context is used
-    pub context_assigner: Option<ContextAssigner<P>>,
+    pub context_assigner: ContextAssigner<P>,
     pub language_mapper: Option<AssetPathToLanguageMapper>,
 
     /// initializers for the contexts, run when loading the script
     pub context_initializers: Vec<ContextInitializer<P>>,
     /// initializers for the contexts run every time before handling events
     pub context_pre_handling_initializers: Vec<ContextPreHandlingInitializer<P>>,
-}
-
-impl<P: IntoScriptPluginParams> Default for ScriptingPlugin<P>
-where
-    P::R: Default,
-{
-    fn default() -> Self {
-        Self {
-            runtime_settings: Default::default(),
-            callback_handler: Default::default(),
-            context_builder: Default::default(),
-            context_assigner: Default::default(),
-            language_mapper: Default::default(),
-            context_initializers: Default::default(),
-            context_pre_handling_initializers: Default::default(),
-        }
-    }
 }
 
 impl<P: IntoScriptPluginParams> Plugin for ScriptingPlugin<P> {
@@ -96,9 +87,9 @@ impl<P: IntoScriptPluginParams> Plugin for ScriptingPlugin<P> {
             })
             .insert_resource::<ContextLoadingSettings<P>>(ContextLoadingSettings {
                 loader: self.context_builder.clone(),
-                assigner: Some(self.context_assigner.clone().unwrap_or_default()),
-                context_initializers: vec![],
-                context_pre_handling_initializers: vec![],
+                assigner: self.context_assigner.clone(),
+                context_initializers: self.context_initializers.clone(),
+                context_pre_handling_initializers: self.context_pre_handling_initializers.clone(),
             });
 
         register_script_plugin_systems::<P>(app);
@@ -113,14 +104,6 @@ impl<P: IntoScriptPluginParams> Plugin for ScriptingPlugin<P> {
         }
 
         register_types(app);
-
-        for initializer in self.context_initializers.iter() {
-            app.add_context_initializer::<P>(*initializer);
-        }
-
-        for initializer in self.context_pre_handling_initializers.iter() {
-            app.add_context_pre_handling_initializer::<P>(*initializer);
-        }
     }
 }
 
@@ -170,9 +153,7 @@ fn once_per_app_init(app: &mut App) {
     app.add_event::<ScriptErrorEvent>()
         .add_event::<ScriptCallbackEvent>()
         .init_resource::<AppReflectAllocator>()
-        .init_resource::<ScriptAssetSettings>()
         .init_resource::<Scripts>()
-        .init_resource::<ScriptMetadataStore>()
         .init_asset::<ScriptAsset>()
         .init_resource::<AppScriptFunctionRegistry>()
         .register_asset_loader(ScriptAssetLoader {
@@ -182,21 +163,10 @@ fn once_per_app_init(app: &mut App) {
 
     app.add_systems(
         PostUpdate,
-        (
-            (garbage_collector).in_set(ScriptingSystemSet::GarbageCollection),
-            (insert_script_metadata).in_set(ScriptingSystemSet::ScriptMetadataInsertion),
-            (remove_script_metadata).in_set(ScriptingSystemSet::ScriptMetadataRemoval),
-        ),
-    )
-    .configure_sets(
-        PostUpdate,
-        (
-            ScriptingSystemSet::ScriptMetadataInsertion.after(bevy::asset::TrackAssets),
-            ScriptingSystemSet::ScriptCommandDispatch
-                .after(ScriptingSystemSet::ScriptMetadataInsertion)
-                .before(ScriptingSystemSet::ScriptMetadataRemoval),
-        ),
+        ((garbage_collector).in_set(ScriptingSystemSet::GarbageCollection),),
     );
+
+    configure_asset_systems(app);
 }
 
 /// Systems registered per-language
@@ -204,11 +174,9 @@ fn register_script_plugin_systems<P: IntoScriptPluginParams>(app: &mut App) {
     app.add_systems(
         PostStartup,
         (initialize_runtime::<P>).in_set(ScriptingSystemSet::RuntimeInitialization),
-    )
-    .add_systems(
-        PostUpdate,
-        ((sync_script_data::<P>).in_set(ScriptingSystemSet::ScriptCommandDispatch),),
     );
+
+    configure_asset_systems_for_plugin::<P>(app);
 }
 
 /// Register all types that need to be accessed via reflection
@@ -238,50 +206,6 @@ impl AddRuntimeInitializer for App {
             .resource_mut::<RuntimeSettings<P>>()
             .as_mut()
             .initializers
-            .push(initializer);
-        self
-    }
-}
-
-pub trait AddContextInitializer {
-    fn add_context_initializer<P: IntoScriptPluginParams>(
-        &mut self,
-        initializer: ContextInitializer<P>,
-    ) -> &mut Self;
-}
-
-impl AddContextInitializer for App {
-    fn add_context_initializer<P: IntoScriptPluginParams>(
-        &mut self,
-        initializer: ContextInitializer<P>,
-    ) -> &mut Self {
-        self.world_mut()
-            .init_resource::<ContextLoadingSettings<P>>();
-        self.world_mut()
-            .resource_mut::<ContextLoadingSettings<P>>()
-            .as_mut()
-            .context_initializers
-            .push(initializer);
-        self
-    }
-}
-
-pub trait AddContextPreHandlingInitializer {
-    fn add_context_pre_handling_initializer<P: IntoScriptPluginParams>(
-        &mut self,
-        initializer: ContextPreHandlingInitializer<P>,
-    ) -> &mut Self;
-}
-
-impl AddContextPreHandlingInitializer for App {
-    fn add_context_pre_handling_initializer<P: IntoScriptPluginParams>(
-        &mut self,
-        initializer: ContextPreHandlingInitializer<P>,
-    ) -> &mut Self {
-        self.world_mut()
-            .resource_mut::<ContextLoadingSettings<P>>()
-            .as_mut()
-            .context_pre_handling_initializers
             .push(initializer);
         self
     }
@@ -325,49 +249,5 @@ impl<D: DocumentationFragment> StoreDocumentation<D> for App {
         }
 
         top_fragment.gen_docs()
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use asset::ScriptMetadataStore;
-
-    use super::*;
-
-    #[test]
-    fn test_default_scripting_plugin_initializes_all_resources_correctly() {
-        let mut app = App::new();
-
-        #[derive(Default, Clone)]
-        struct C;
-        #[derive(Default, Clone)]
-        struct R;
-
-        struct Plugin;
-
-        impl IntoScriptPluginParams for Plugin {
-            type C = C;
-            type R = R;
-            const LANGUAGE: Language = Language::Unknown;
-
-            fn build_runtime() -> Self::R {
-                R
-            }
-        }
-
-        app.add_plugins(AssetPlugin::default());
-        app.add_plugins(ScriptingPlugin::<Plugin>::default());
-
-        assert!(app.world().contains_resource::<Scripts>());
-        assert!(app.world().contains_resource::<AppTypeRegistry>());
-        assert!(app.world().contains_resource::<ScriptAssetSettings>());
-        assert!(app.world().contains_resource::<RuntimeSettings<Plugin>>());
-        assert!(app.world().contains_resource::<CallbackSettings<Plugin>>());
-        assert!(app
-            .world()
-            .contains_resource::<ContextLoadingSettings<Plugin>>());
-        assert!(app.world().contains_non_send::<RuntimeContainer<Plugin>>());
-        assert!(app.world().contains_non_send::<ScriptContexts<Plugin>>());
-        assert!(app.world().contains_resource::<ScriptMetadataStore>());
     }
 }
